LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.2_Variable_Statement/S12.2_A8_T7-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Parse error
S12.2_A8_T7

PASS Expected parsing failure

TEST COMPLETE
